chore: Update self sign certificate names
The code changes update the names of the self sign certificates used for VPN. The root certificate name is changed to "p2srootcertv7" and the client certificate name is changed to "p2sclientcertv7". This update ensures that the certificates are up to date and aligned with the current version of the VPN.

Note: The recent repository commits indicate that the developer has been working on this feature.